move site and social link components to own folder
