+++ src/components/Links/index.ts
@@ -1,0 +1,2 @@
+export { default as SiteLinks } from './siteLinks';
+export { default as SocialLinks } from './socialLinks';
